+ -Wconversion

+++ backend_x11/graphics_system/native/x11_icon.cpp
@@ -53,7 +53,7 @@ X11Icon::X11Icon(
 
         for(int y = 0; y< height; y++) {
             for(int x = 0; x < width; x++) {
-                uint32_t pixel = argb_data[(y * width) + x];
+                uint32_t pixel = static_cast<uint32_t>(argb_data[(y * width) + x]);
 
                 unsigned int alpha = (pixel & 0xFF000000) >> 24;
                 unsigned int red_in = (pixel & 0xFF0000) >> 16;
